Changes:
- split edge cases tests
