+++ tests/test_direct_responses.py
@@ -1,18 +1,11 @@
 import os
-from typing import Optional
 
-from pydantic import BaseModel
-
-from esmerald import Esmerald, Form, Redirect, Request, Template
+from esmerald import Esmerald, Redirect, Request, Template
 from esmerald.config.template import TemplateConfig
 from esmerald.responses.base import RedirectResponse
 from esmerald.routing.gateways import Gateway
-from esmerald.routing.handlers import get, route
+from esmerald.routing.handlers import get
 from esmerald.testclient import EsmeraldTestClient
-
-
-class Model(BaseModel):
-    id: str
 
 
 def test_return_response_container(template_dir):
@@ -56,17 +49,3 @@ def test_return_response(template_dir):
     client = EsmeraldTestClient(app)
     response = client.get("/", follow_redirects=False)
     assert response.status_code == 301
-
-
-def test_get_and_post():
-    @route(methods=["GET", "POST"])
-    async def start(request: Request, data: Optional[Model] = Form()) -> bytes:
-        return b"hello world"
-
-    app = Esmerald(
-        debug=True,
-        routes=[Gateway("/", handler=start)],
-    )
-    client = EsmeraldTestClient(app)
-    response = client.get("/")
-    assert response.status_code == 200
